Finished making nested parentheses work
I imagine this is a bit like what the aftermath of childbirth feels like. It was so painful, but now I have this child and I love it so much and I feel like I brought a great thing into the world.

Except that my child turns out to be a mean person that brings no joy to anyone. Do I still love it? Was childbirth really worth it?